+++ docs/src/demos/components/Inset/index.ts
@@ -1,0 +1,3 @@
+export * from './inset-basic.demo';
+export * from './inset-equal.demo';
+export * from './inset-hv.demo';