+++ docs/src/demos/components/Inset/inset-equal.demo.tsx
@@ -1,0 +1,9 @@
+import { Inset } from '@marigold/components';
+
+export const InsetEqualSpaceDemo = () => (
+  <Inset space="large">
+    Marigold has oblong and lanceolate leaves with whole margins. Some varieties
+    of marigold have leaves with toothed edges. Leaves are spirally arranged on
+    the branches.
+  </Inset>
+);